Refactor: Move player files into player mod
